fix: refactor integration statistic
