+++ type/statisticData.ts
@@ -1,5 +1,5 @@
 export type StatisticData = {
-    active_user: number;
-    user_claim: number;
-    total_user: number;
+    activeUser: number;
+    userClaim: number;
+    totalUser: number;
   }
