+++ type/statisticResponse.ts
@@ -3,5 +3,5 @@ import type { StatisticData } from "./statisticData";
 export type StatisticResponse = {
     data: StatisticData;
     message: string;
-    status_code: number;
+    statusCode: number;
   }
