+++ type/statisticState.ts
@@ -1,0 +1,6 @@
+import type { StatisticData } from "~/type/statisticData";
+
+export type StatisticState = {
+  statistics: StatisticData;
+  isLoading: boolean;
+};
